* Quicktime edit lists
git-svn-id: svn://svn.code.sf.net/p/gmerlin/code/trunk@1546 214d8ac5-8513-4f59-80f8-2f39b674a24a

+++ gmerlin_avdecoder/include/qt.h
@@ -706,7 +706,7 @@ typedef struct
     {
     uint32_t duration;
     uint32_t media_time;
-    float media_rate;
+    uint32_t media_rate;
     } * table;
   } qt_elst_t;
